# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feature-requests:
    post:
      operationId: createFeatureRequest
      tags: [feature-requests]
      summary: Send a feature request to the HI-DO-RI team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureRequestInput"
      responses:
        "201":
          description: Feature request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureRequestResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Notification could not be delivered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feature-requests/development:
    get:
      operationId: listFeatureRequestDevelopment
      tags: [feature-requests]
      summary: List feature requests for the operator development board
      responses:
        "200":
          description: Development board
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureRequestDevelopmentItem"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Operator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feature-requests/{id}/development-status:
    post:
      operationId: updateFeatureRequestDevelopmentStatus
      tags: [feature-requests]
      summary: Record a milestone and notify Slack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureRequestDevelopmentStatusInput"
      responses:
        "200":
          description: Milestone recorded and notified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureRequestDevelopmentItem"
        "400":
          description: Invalid update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Operator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Milestone already recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Slack notification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    FeatureRequestInput:
      type: object
      required: [name, request]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        contact:
          type: string
          maxLength: 160
        request:
          type: string
          minLength: 10
          maxLength: 2000
        useCase:
          type: string
          maxLength: 1000
    FeatureRequestResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    FeatureRequestDevelopmentStatus:
      type: string
      enum: [received, approved, in_progress, validating, preview_ready, completed]
    FeatureRequestDevelopmentStatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/FeatureRequestDevelopmentStatus"
        note:
          type: string
          maxLength: 500
    FeatureRequestDevelopmentItem:
      type: object
      required: [id, name, request, notificationStatus, developmentStatus, createdAt, developmentStatusUpdatedAt]
      properties:
        id:
          type: number
        name:
          type: string
        request:
          type: string
        useCase:
          type: ["string", "null"]
        notificationStatus:
          type: string
        developmentStatus:
          $ref: "#/components/schemas/FeatureRequestDevelopmentStatus"
        developmentNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        developmentStatusUpdatedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [google_calendar_reconnect_required]