# lib/api-zod/src/generated/types/featureRequestDevelopmentStatusInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureRequestDevelopmentStatus } from './featureRequestDevelopmentStatus';

export interface FeatureRequestDevelopmentStatusInput {
  status: FeatureRequestDevelopmentStatus;
  /** @maxLength 500 */
  note?: string;
}